refactored project file structure
