+++ src/App.js
@@ -1,11 +1,11 @@
-import React, { Component } from 'react'
-import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
+import { Component } from 'react';
+import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 
-import Login from './Login'
-import LoginSuccess from './LoginSuccess'
-import Profile from './Profile'
-import Home from './Home'
-import Signup from './Signup'
+import Home from './components/home/Home';
+import Login from './components/login/Login';
+import LoginSuccess from './components/login/LoginSuccess';
+import Signup from './components/login/Signup';
+import Profile from './components/profile/Profile';
 
 export default class App extends Component {
   render() {
@@ -21,6 +21,6 @@ export default class App extends Component {
           </Routes>
         </div>
       </Router>
-    )
+    );
   }
 }
