+++ src/components/home/Post.js
@@ -1,5 +1,3 @@
-import React from 'react'
-
 export default function Post() {
-  return <div></div>
+  return <div></div>;
 }
