+++ src/components/home/Stories.js
@@ -1,11 +1,9 @@
-import React from 'react'
+import Story from './Story';
 
-import Story from './Story'
-
-import profile from '../../assets/images/profile.jpg'
-import profile2 from '../../assets/images/profile2.jpg'
-import profile3 from '../../assets/images/profile3.jpg'
-import profile4 from '../../assets/images/profile4.jpg'
+import profile from '../../assets/images/profile.jpg';
+import profile2 from '../../assets/images/profile2.jpg';
+import profile3 from '../../assets/images/profile3.jpg';
+import profile4 from '../../assets/images/profile4.jpg';
 
 export default function Stories(props) {
   return (
@@ -15,5 +13,5 @@ export default function Stories(props) {
       <Story image={profile3} username={'worthless.safi'} />
       <Story image={profile4} username={'dawnmariel'} />
     </div>
-  )
+  );
 }
