+++ src/components/profile/ProfileMediaItem.js
@@ -1,9 +1,7 @@
-import React from 'react'
-
 export default function ProfileMediaItem(props) {
   return (
     <div className='profile-media-item'>
       <img src={props.preview} alt='' />
     </div>
-  )
+  );
 }
